requirements.txt
aiohttp==3.8.1
pycryptodome==3.14.1
